adb: Fix adb_iovec structure for 64-bit Windows
This changes the type of iov_len member in adb_iovec structure
to int (4 bytes). It now matches with the type of len member
of WSABUF structure defined in ws2def.h file.

This fixes the following assertion in 64-bit Windows platform:

adb/sysdeps_win32.cpp:720:49: error: static assertion failed
  720 | static_assert(SIZEOF_MEMBER(adb_iovec, iov_len) == SIZEOF_MEMBER(WSABUF, len), "");
adb/sysdeps_win32.cpp:720:49: note: the comparison reduces to '(8 == 4)'

+++ sysdeps/uio.h
@@ -23,8 +23,9 @@
 #if defined(_WIN32)
 
 // Layout of this struct must match struct WSABUF (verified via static assert in sysdeps_win32.cpp)
+// ws2def.h: typedef struct _WSABUF { ULONG len; CHAR *buf; } WSABUF, *LPWSABUF;
 struct adb_iovec {
-    size_t iov_len;
+    int iov_len;
     void* iov_base;
 };
